Replace the Accounts impl by a TransactionLog
The Accounts were a premature optimisation: i duplicated events for
every account in the bank, so that i could reduce over one account. This
in turn resulted in a consistency violation: I had to do two writes.
This is all solved by building a proper projection in the read side of
the bank.